requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.40.0
python-telegram-bot>=21.0
python-multipart>=0.0.9
requests>=2.31.0
yt-dlp>=2024.1.0
